% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biostrings_alignment.R
\name{calc_biostrings_alignment}
\alias{calc_biostrings_alignment}
\title{Calculate dissimilarity scores based on Smith-Waterman or Needleman-Wunsch}
\usage{
calc_biostrings_alignment(
  X,
  ncpus = 1,
  seqtype = c("aa", "dna", "rna"),
  par.list = list(),
  vrb = TRUE
)
}
\arguments{
\item{X}{data frame with two fields, `IDs` (with sequence ids) and `SEQs`
(containing strings with the sequences to be aligned)}

\item{ncpus}{number of cores to use}

\item{seqtype}{type of sequence being aligned. Accepts "aa", "dna" or "rna".}

\item{par.list}{list object with parameters to be passed to
[Biostrings::pairwiseAlignment()]. See `Details`.}

\item{vrb}{logical flag: should progress be printed to console?}
}
\value{
list object with two elements: `scores` (matrix of
local or global alignment scores) and `diss_matrix` (dissimilarity matrix)
}
\description{
This function invokes [Biostrings::pairwiseAlignment()] to calculate the
normalised dissimilarity scores between all pairs of sequences in a
data frame.
}
