% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diamond_alignment.R
\name{calc_diamond_alignment}
\alias{calc_diamond_alignment}
\title{Calculate dissimilarity scores based on DIAMOND-BLAST alignment}
\usage{
calc_diamond_alignment(
  X = NULL,
  seqfile = NULL,
  ncpus = 1,
  seqtype = c("aa", "dna", "rna"),
  par.list = c("--sensitive", "--matrix BLOSUM62", "--gapopen 11", "--gapextend 1",
    "--block-size 0.5", "--quiet"),
  diamond.path = "./",
  min.hit = 8,
  cleanup = TRUE,
  vrb = TRUE
)
}
\arguments{
\item{X}{data frame with two fields, `IDs` (with sequence ids) and `SEQs`
(containing strings with the sequences to be aligned). Ignored if a file
path is provided in `seqfile`.}

\item{seqfile}{FASTA file containing the sequences. If `NULL` then sequences
must be provided as a data frame in `X`.}

\item{ncpus}{number of cores to use}

\item{seqtype}{type of sequence being aligned. Accepts "aa", "dna" or "rna".}

\item{par.list}{list object with parameters to be passed to
DIAMOND. See `Details`.}

\item{diamond.path}{path to the folder where DIAMOND can be executed
(e.g., where the executable file is located). This is also the folder where
the DIAMOND files will be (temporarily) saved.}

\item{min.hit}{shortest allowed length of alignment hit.}

\item{cleanup}{logical flag: should files created by DIAMOND be deleted at
the end?}

\item{vrb}{logical flag: should progress be printed to console? Note that
this does not control the echoing of DIAMOND - for that, add option
`"--quiet"` to `par.list`.}
}
\value{
list object with two elements: `scores` (matrix of
local or global alignment scores) and `diss_matrix` (dissimilarity matrix)
}
\description{
This function invokes DIAMOND (needs to be installed externally) to
 calculate the
normalised dissimilarity scores between all pairs of sequences in a
data frame.
}
\section{Installing DIAMOND}{

This routine requires DIAMOND to be available in the system. Please refer to
[https://github.com/bbuchfink/diamond/wiki/2.-Installation](https://github.com/bbuchfink/diamond/wiki/2.-Installation) for installation details.
}

\section{Details}{

Parameter `par.list` should contain a character vector or a list of options
that the user wants to pass to DIAMOND.
By default, this list will be populated with:

\itemize{
 \item `--ultra-sensitive` (the most sensitive setting)
 \item `--block-size 0.5`   (to limit memory usage to about 3GB, which should enable it to run in most personal machines)
 \item `--matrix BLOSUM62` (DIAMOND standard)
 \item `--gapopen 11` (DIAMOND standard)
 \item `--gapextend 1` (DIAMOND standard)
}

Additionally, this routine will translate the value of `ncpus`
to the DIAMOND option `--threads ncpus`.

The user can pass any desired parameters for DIAMOND, except `--outfmt`
(which is always set to 6, the BLAST tabular format with
the default output fields defined in DIAMOND:
"qseqid", "sseqid", "pident", "length",  "mismatch", "gapopen",
"qstart", "qend", "sstart", "send", "evalue", "bitscore")
}

\section{Dissimilarity value}{

The dissimilarity values returned by this function are calculated as:

diss(a, b) = 1 - max(pident(a, b)) / 100,

that is, they are based on the largest percent identity (as returned by
DIAMOND) identified between two sequences "a" and "b". Note that only hits
longer than `min.hit` are considered.
}

