% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_split_homogeneity.R
\name{calc_split_homogeneity}
\alias{calc_split_homogeneity}
\title{Calculate split homogeneity for all splits given an allocation matrix.}
\usage{
calc_split_homogeneity(X, C)
}
\arguments{
\item{X}{matrix of binary allocation variables. Each position \eqn{x_{ki}}
indicates the allocation or not of the i-th group to the k-th split. Note
that each group must be allocated to a single split, i.e.,
colSums(X) must be a vector of ones.}

\item{C}{matrix of class counts per group. Commonly calculated using
[consolidate_class_counts()]. Each position \eqn{c_{ij}} contains
the number of examples of the j-th Class contained in the i-th group.}
}
\value{
A vector containing the normalized homogeneity scores of
each split.
}
\description{
Calculate split homogeneity for all splits given an allocation matrix.
}
\examples{

library(moses)

C <- matrix(c(10, 10, 1, 5, 1, 4, 5, 15, 0),
            nrow = 3, byrow = TRUE)

X <- matrix(c(0, 0, 1, 1, 1, 0),
           nrow = 2, byrow = TRUE)

calc_split_homogeneity(X, C)

}
