% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_balance_deviation.R
\name{calculate_balance_deviation}
\alias{calculate_balance_deviation}
\title{Calculate balance deviation for all splits given an allocation matrix.}
\usage{
calculate_balance_deviation(X, C)
}
\arguments{
\item{X}{matrix of binary allocation variables. Each position `x_{ki}`
indicates the allocation or not of the i-th group to the k-th split. Note
that each group must be allocated to a single split, i.e.,
colSums(X) must be a vector of ones.}

\item{C}{matrix of class counts per group. Commonly calculated using
[consolidate_class_counts()]. Each position `c_{ij}` contains
the number of examples of the j-th Class contained in the i-th group.}
}
\value{
A matrix containing the normalized balance deviation scores of each
class on each split (splits as rows, classes as columns).
}
\description{
Calculate balance deviation for all splits given an allocation matrix.
}
\examples{
\dontrun{
library(moses)

# Build example with 6 groups, 3 classes, 2 splits

C <- matrix(c(1, 0, 0,
              3, 5, 0,
              4, 5, 6,
              4, 5, 6,
              4, 5, 6,
              4, 5, 6),
              ncol = 3, byrow = TRUE)

X = matrix(0,
           nrow = 2,
           ncol = nrow(C))
X[1, 1:2] <- 1
X[2, 3:6] <- 1

calculate_balance_deviation(X, C)
}

}
