% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_split_homogeneity.R
\name{calculate_split_homogeneity}
\alias{calculate_split_homogeneity}
\title{Calculate split homogeneity for all splits given an allocation matrix.}
\usage{
calculate_split_homogeneity(X, C)
}
\arguments{
\item{X}{matrix of binary allocation variables. Each position `x_{ki}`
indicates the allocation or not of the i-th group to the k-th split. Note
that each group must be allocated to a single split, i.e.,
colSums(X) must be a vector of ones.}
}
\value{
@return A vector containing the normalized homogeneity scores of
each split.
}
\description{
Calculate split homogeneity for all splits given an allocation matrix.
}
\examples{
\dontrun{
library(moses)

# Build example with 3 splits, 6 groups

X = matrix(0, nrow = 3, ncol = 6)
X[1, 1] <- 1
X[2, 2:3] <- 1
X[3, 4:6] <- 1

calculate_split_homogeneity(X)
}

}
