% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cdhit_clusters.R
\name{extract_cdhit_clusters}
\alias{extract_cdhit_clusters}
\title{Extract protein sequence clusters using CDHIT}
\usage{
extract_cdhit_clusters(
  X = NULL,
  seqfile = NULL,
  diss_threshold = 0.3,
  vrb = TRUE,
  ncpus = 1,
  par.list = list(kmerSize = NULL, min_length = 8, s = , G = 1)
)
}
\arguments{
\item{X}{data frame with two fields, `IDs` (with sequence ids) and `SEQs`
(containing strings with the sequences to be aligned). Ignored if a file
path is provided in `seqfile`.}

\item{seqfile}{FASTA file containing the sequences. If `NULL` then sequences
must be provided as a data frame in `X`.}

\item{diss_threshold}{the desired dissimilarity cutting threshold for
attributing clusters. Passed down to [CellaRepertorium::cdhit()] as
parameter `identity = 1 - diss_treshold`.}

\item{vrb}{logical flag: should progress be printed to console? Passed down
to [CellaRepertorium::cdhit()] as parameter `showProgress = vrb`.}

\item{ncpus}{number of cores to use. Passed down
to [CellaRepertorium::cdhit()] as parameter `T = ncpus`.}

\item{par.list}{further parameters to be passed down to
[CellaRepertorium::cdhit()]. Please check that function for details.}
}
\value{
list object with two elements: `cl.df` (A data frame returned by
CDHIT) and `clusters` (a data frame
with cluster memberships at the dissimilarity level `diss_threshold`).
}
\description{
This function extract protein sequence clusters using CDHIT
(<https://doi.org/10.1093/bioinformatics/btl158>), based on the R
function [CellaRepertorium::cdhit()]. Please refer to that function for
details.
}
