% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clusters_cdhit.R
\name{extract_clusters_cdhit}
\alias{extract_clusters_cdhit}
\title{Extract protein sequence clusters using CDHIT}
\usage{
extract_clusters_cdhit(
  X = NULL,
  seqfile = NULL,
  diss_threshold = 0.3,
  vrb = TRUE,
  ncpus = 1,
  par.list = list(kmerSize = NULL, min_length = 8, s = 1, G = 1)
)
}
\arguments{
\item{X}{data frame with two fields, `IDs` (with sequence ids) and `SEQs`
(containing strings with the sequences to be aligned). Ignored if a file
path is provided in `seqfile`.}

\item{seqfile}{FASTA file containing the sequences. If `NULL` then sequences
must be provided as a data frame in `X`.}

\item{diss_threshold}{the desired dissimilarity cutting threshold for
attributing clusters. Passed down to [CellaRepertorium::cdhit()] as
parameter `identity = 1 - diss_treshold`.}

\item{vrb}{logical flag: should progress be printed to console? Passed down
to [CellaRepertorium::cdhit()] as parameter `showProgress = vrb`.}

\item{ncpus}{number of cores to use. Passed down
to [CellaRepertorium::cdhit()] as parameter `T = ncpus`.}

\item{par.list}{further parameters to be passed down to
[CellaRepertorium::cdhit()]. Please check that function for details.}
}
\value{
list object with two elements: `cl.df` (A data frame returned by
CDHIT) and `clusters` (a data frame
with cluster memberships at the dissimilarity level `diss_threshold`).
}
\description{
This function extract protein sequence clusters using CDHIT
(<https://doi.org/10.1093/bioinformatics/btl158>), based on the R
function [CellaRepertorium::cdhit()]. Please refer to that function for
details.
}
\examples{

library(moses)

# load data file
fpath <- system.file("extdata", "some_prots.rds", package="moses")
X <- readRDS(fpath)
mycl <- extract_clusters_cdhit(X)

using a fasta file directly
fpath <- system.file("diamond", "bfv_proteins.fa", package="moses")
mycl <- extract_clusters_cdhit(seqfile = fpath, diss_threshold = 0.2)

}
