% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSES_snp_meth.R
\name{MOSES_snp_meth}
\alias{MOSES_snp_meth}
\title{Train gene expression prediction model and conduct TWAS using customized training data}
\usage{
MOSES_snp_meth(
  example = TRUE,
  train.meth.file,
  train.exp.file,
  test.meth.file,
  genotype.file.list.file,
  TWAS = TRUE,
  pheno.file,
  phenotype,
  confounder,
  output.file.path,
  core.num = 1
)
}
\arguments{
\item{example}{An indicator to input example datasets. Default is FALSE.}

\item{train.meth.file}{A file containing methylation matrix as training data.}

\item{train.exp.file}{A file containing gene expression matrix as training data.}

\item{test.meth.file}{A file containing methylation matrix as test data.}

\item{genotype.file.list.file}{A txt file containing three columns named as "chr", "train.snp.file", and "test.snp.file". The "chr" column contains numeric chromosome number (e.g. 1, 2, ..., 22); the "train.snp.file" contains the directory of the training set genotype file from the corresponding chr; the "train.snp.file" contains the directory of the test set genotype file from the corresponding chr}

\item{TWAS}{An indicator to perform TWAS analysis. Default is TRUE.}

\item{pheno.file}{A file containing phenotype information.}

\item{confounder}{Names of variables to be adjusted in TWAS.}

\item{output.file.path}{The path/directory to store output files.}

\item{core.num}{The number of cores to be used. Default is 1.}

\item{pheno}{The name of outcome of interest in TWAS.}
}
\value{
Two output files in the folder specified by output.file.path: 1) prediction.Rdata is a matrix containing gene expression value predicted using train.meth.rda and train.exp.rda 2) TWAS.result.txt is a matrix containing TWAS results.
}
\description{
Train gene expression prediction model and conduct TWAS using customized training data
}
\examples{
MOSES_snp_meth(example = F, train.meth.file = "/ix/ksoyeon/YQ/code/MethylTWAS/data/train.meth.rda",test.meth.file = "/ix/ksoyeon/YQ/code/MethylTWAS/data/test.meth.rda",train.exp.file = "/ix/ksoyeon/YQ/code/MethylTWAS/data/train.exp.rda",genotype.file.list="/ix/ksoyeon/YQ/code/MethylTWAS/data/genotype.file.list.txt",pheno.file = "/ix/ksoyeon/YQ/code/MethylTWAS/data/pheno.rda",output.file.path = "/ix/ksoyeon/YQ/results/test/",TWAS = T,phenotype = "cc_new",confounder = "gender, age")
MOSES_snp_meth(example = T, output.file.path = "/ix/ksoyeon/YQ/results/test/",TWAS = T,phenotype = "cc_new",confounder = "gender, age")
}
