% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_objective_function.R
\name{scalar_objective_function}
\alias{scalar_objective_function}
\title{Calculate the scalarised objective function for data splitting}
\usage{
scalar_objective_function(
  alloc_vector,
  class_counts,
  target_proportions,
  weights = 1,
  rho = 0.01,
  class_balance = NULL
)
}
\arguments{
\item{alloc_vector}{vector containing the allocation information
for each cluster of observations. Must be a vector of
length `nrow(class_counts)`. Each position __i__ should contain
an integer between 1 and __nsplits__ indicating the split number to
which the __i__-th cluster is allocated.}

\item{class_counts}{data frame with one row per cluster.
Must contain a column Cluster` (with the cluster number/ID) plus
additional columns with counts of occurrences for each unique class
and columns with the class proportions in each cluster, plus a
column with the total cluster size.}

\item{weights}{vector of weights for function aggregation. Must be of
length 3 and add to 1. All weights must be non-negative. If a
scalar value is passed then the weights are taken as equal for all
objectives. Weights are scaled proportionally if they don't add to 1.}

\item{rho}{small positive value for augmented Tchebycheff scalarisation.}

\item{class_balance}{A data frame with the proportions of each class in
the data. Calculated internally if `NULL`.}
}
\description{
This function calculates the objective function for the data splitting
optimisation. See section **Scalarisation details** for a full description.
}
\section{Scalarisation details}{

This function calculates a scalarised objective function that is minimised
as part of the definition of the optimal data splits for sequence data
under a clustering/grouping structure, with potential data imbalance and
with an arbitrary number of classes.
The three objective functions (all of which are set to be minimised, i.e.,
smaller = better) are:
\itemize{
   \item f1 (*Size deviation*), which measures how much the resulting splits
   (given a tentative allocation) deviate from the desired proportion.
   \item f2 (*Balance deviation*), which measures how much the class balance of
   the splits deviates from the overall class balance of the full dataset.
   \item f3 (*Homogeneity score*), which quantifies how homogeneous (using the
   number of different clusters/groups in each split as a measure of
   diversity) the splits are.
}

These three objective functions are aggregated using the augmented
Tchebycheff scalarisation function,

$$
f_{AT} = \max_\ell\left{w_\ell f_\ell(x)\right} + \rho\sum_{\ell}w_\ell f_\ell(x)
$$

where $w$ represent the weights provided by the user, and $\rho$ is a small
positive value.
}

