% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biostrings_alignment.R,
%   R/calc_diamond_alignment.R
\name{calc_biostrings_alignment}
\alias{calc_biostrings_alignment}
\title{Calculate dissimilarity scores based on Smith-Waterman or Needleman-Wunsch}
\usage{
calc_biostrings_alignment(
  X,
  ncpus = 1,
  seqtype = c("aa", "dna", "rna"),
  par.list = list(),
  vrb = TRUE
)

calc_biostrings_alignment(
  X,
  ncpus = 1,
  seqtype = c("aa", "dna", "rna"),
  par.list = list(),
  vrb = TRUE
)
}
\arguments{
\item{X}{data frame with two fields, `IDs` (with sequence ids) and `SEQs`
(containing strings with the sequences to be aligned)}

\item{ncpus}{number of cores to use}

\item{seqtype}{type of sequence being aligned. Accepts "aa", "dna" or "rna".}

\item{par.list}{list object with parameters to be passed to
DIAMOND. See `Details`.}

\item{vrb}{logical flag: should progress be printed to console?}

\item{diamond.path}{path to the folder where DIAMOND can be executed
(e.g., where the executable file is located).}
}
\value{
list object with two elements: `scores` (matrix of
local or global alignment scores) and `diss_matrix` (dissimilarity matrix)

list object with two elements: `scores` (matrix of
local or global alignment scores) and `diss_matrix` (dissimilarity matrix)
}
\description{
This function invokes [Biostrings::pairwiseAlignment()] to calculate the
normalised dissimilarity scores between all pairs of sequences in a
data frame.

This function invokes DIAMOND (needs to be installed externally) to
 calculate the
normalised dissimilarity scores between all pairs of sequences in a
data frame.
}
\section{Details}{

Parameter `par.list` should contain the list of parameters that the user
wants to pass down to [Biostrings::pairwiseAlignment()]. By default,
this list will be populated with:

\itemize{
 \item `type = "local"`
 \item `substitutionMatrix = "BLOSUM62"`
 \item `gapOpening = 10`
 \item `gapExtension = 4`
}

The user can overwrite these values and add any other parameter accepted by
[Biostrings::pairwiseAlignment()] _except_ ` scoreOnly` (which is always set
to `TRUE`).


Parameter `par.list` should contain a character vector or a list of options
that the user wants to pass to DIAMOND.
By default, this list will be populated with:

\itemize{
 \item `"--ultra-sensitive"`
 \item `--matrix BLOSUM62` (DIAMOND standard)
 \item `--gapopen 11` (DIAMOND standard)
 \item `--gapextend 1` (DIAMOND standard)
}

Additionally, this routine will translate the value of `ncpus`
to the DIAMOND option `--threads ncpus`. Similarly, setting `vrb` to
`TRUE` will be translated as option `--quiet` for DIAMOND.

The user can pass any desired parameters for DIAMOND, except `--outfmt`
(which is always set to 6, the BLAST tabular format with
the default output fields defined in DIAMOND:
"qseqid", "sseqid", "pident", "length",  "mismatch", "gapopen",
"qstart", "qend", "sstart", "send", "evalue", "bitscore")
}

\section{Installing DIAMOND}{

This routine requires DIAMOND to be available in the system. Please refer to
[https://github.com/bbuchfink/diamond/wiki/2.-Installation](https://github.com/bbuchfink/diamond/wiki/2.-Installation) for installation details.
}

