% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biostrings_alignment.R
\name{calc_biostrings_alignment}
\alias{calc_biostrings_alignment}
\title{Calculate dissimilarity scores based on Smith-Waterman or Needleman-Wunsch}
\usage{
calc_biostrings_alignment(
  X = NULL,
  seqfile = NULL,
  ncpus = 1,
  seqtype = c("aa", "dna", "rna"),
  par.list = list(type = "local", substitutionMatrix = "BLOSUM62", gapOpening = 10,
    gapExtension = 4),
  vrb = TRUE
)
}
\arguments{
\item{X}{data frame with two fields, `IDs` (with sequence ids) and `SEQs`
(containing strings with the sequences to be aligned). Ignored if a file
path is provided in `seqfile`.}

\item{seqfile}{FASTA file containing the sequences. If `NULL` then sequences
must be provided as a data frame in `X`.}

\item{ncpus}{number of cores to use}

\item{seqtype}{type of sequence being aligned. Accepts "aa", "dna" or "rna".}

\item{par.list}{list object with parameters to be passed to
[Biostrings::pairwiseAlignment()]. See `Details`.
A `NULL` par.list is translated to the default values.}

\item{vrb}{logical flag: should progress be printed to console?}
}
\value{
list object with two elements: `scores` (matrix of
local or global alignment scores) and `diss_matrix` (dissimilarity matrix)
}
\description{
This function invokes [Biostrings::pairwiseAlignment()] to calculate the
normalised dissimilarity scores between all pairs of sequences in a
data frame.
}
\section{Details}{

Parameter `par.list` should contain the list of parameters that the user
wants to pass down to [Biostrings::pairwiseAlignment()]. By default,
this list will be populated with:

\itemize{
 \item `type = "local"`
 \item `substitutionMatrix = "BLOSUM62"`
 \item `gapOpening = 10`
 \item `gapExtension = 4`
}

The user can overwrite these values and add any other parameter accepted by
[Biostrings::pairwiseAlignment()] _except_ ` scoreOnly` (which is always set
to `TRUE`).
}

