% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_split_homogeneity.R
\name{calc_split_homogeneity}
\alias{calc_split_homogeneity}
\title{Calculate split homogeneity for all splits given an allocation matrix.}
\usage{
calc_split_homogeneity(X)
}
\arguments{
\item{X}{matrix of binary allocation variables. Each position `x_{ki}`
indicates the allocation or not of the i-th group to the k-th split. Note
that each group must be allocated to a single split, i.e.,
colSums(X) must be a vector of ones.}
}
\value{
A vector containing the normalized homogeneity scores of
each split.
}
\description{
Calculate split homogeneity for all splits given an allocation matrix.
}
\examples{

library(moses)

X <- matrix(c(1, 0, 0, 0, 0, 0,
              0, 1, 1, 0, 0, 0,
              0, 0, 0, 1, 1, 1),
           nrow = 3, byrow = TRUE)

calc_split_homogeneity(X)

}
