% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_class_counts.R
\name{consolidate_class_counts}
\alias{consolidate_class_counts}
\title{Consolidate ID, cluster and class count data}
\usage{
consolidate_class_counts(clusters, class_counts)
}
\arguments{
\item{clusters}{data frame containing entry IDs (column `ID`) and their
respective cluster identifiers (column `Cluster`).}

\item{class_counts}{data frame containing entry IDs (column `ID` - may
contain repeated values, in which case they are aggregated using the sum of
each class label) and either:
(i) a single column named `Class` with the class label associated with the
respective ID, or (ii) two or more columns (with any arbitrary names
**except** `Class`. `Class.A`, `Class.B` etc. are permitted) each containing
the counts of different classes associated with the respective ID.
Note that in the second case there should be no other columns except `ID` and
the class columns. See "Examples".}
}
\value{
A list object containing:
\itemize{
   \item A data frame with one row per cluster. Contains a column
`   Cluster` (with the cluster number/ID) and additional columns with
    counts of occurrences for each unique class (with
    names taken either from the unique class values in
`   class_counts$Class` or from the column names of `class_counts`),
    plus columns with the class proportions in each cluster and a
    column with the total cluster size.
   \item A data frame with the proportions of each class in the data.
}
}
\description{
This function consolidates the output of [extract_clusters()] or
[extract_clusters_cdhit()] with data related to the count of class examples
to be associated with each entry.
}
\examples{
\dontrun{
library(moses)

# using any fasta file.
x <- calc_seq_dissimilarities(seqfile = "diamond/bfv_proteins.fa",
                              aligner = "SW")
cl <- extract_clusters(x$diss_matrix)

epits <- readRDS("diamond/bfv_peptides.rds")

cc <- consolidate_class_counts(cl$clusters, epits)
}


}
