% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_class_counts.R
\name{consolidate_class_counts}
\alias{consolidate_class_counts}
\title{Consolidate ID, cluster and class count data}
\usage{
consolidate_class_counts(clusters, class_counts)
}
\arguments{
\item{clusters}{data frame containing entry IDs (column `ID`) and their
respective cluster identifiers (column `Cluster`).}

\item{class_counts}{data frame containing entry IDs (column `ID` - may
contain repeated values) and either:
(i) a single column named `Class` with the class label associated with the
respective ID, or (ii) two or more columns (with any arbitrary names
**except** `Class`. `Class.A`, `Class.B` etc. are permitted) each containing
the counts of different classes associated with the respective ID.
Note that in the second case there should be no other columns except `ID` and
the class columns. See "Examples".}
}
\value{
A data frame with one row per cluster, a `Cluster` column and
additional columns with counts of occurrences for each unique class (with
names taken either from the unique class values or from the column names of
`class_counts`).
}
\description{
This function consolidates the output of [extract_clusters()] or
[extract_cdhit_clusters()] with data related to the count of class examples
to be associated with each entry.
}
