% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_class_counts.R
\name{consolidate_class_counts}
\alias{consolidate_class_counts}
\title{Consolidate ID, group and class count data}
\usage{
consolidate_class_counts(groups, class_counts)
}
\arguments{
\item{groups}{data frame containing entry IDs (column `ID`) and their
respective group identifiers (column `Cluster`).}

\item{class_counts}{data frame containing entry IDs (column `ID` - may
contain repeated values, in which case they are aggregated using the sum of
each class label) and either:
(i) a single column named `Class` with the class label associated with the
respective ID, or (ii) two or more columns (with any arbitrary names
**except** `Class`. `Class.A`, `Class.B` etc. are permitted) each containing
the counts of different classes associated with the respective ID.
Note that in the second case there should be no other columns except `ID` and
the class columns. See "Examples".}
}
\value{
A matrix with one row per group, and the counts of each unique class
   (with names taken either from the unique class values in
   ` class_counts$Class` or from the column names of `class_counts`) as
   columns.
}
\description{
This function consolidates the output of [extract_clusters()] or
[extract_clusters_cdhit()] with data related to the count of class examples
to be associated with each entry.
}
\examples{
\dontrun{
library(moses)

fpath1 <- system.file("diamond", "bfv_proteins.fa", package="moses")
fpath2 <- system.file("diamond", "bfv_peptides.rds", package="moses")

# Calculate clusters using sequence data
mycl <- extract_clusters_cdhit(seqfile = fpath1, diss_threshold = 0.2)

# Load data frame with classes
df <- readRDS(fpath2)

# Consolidate class counts
C <- consolidate_class_counts(mycl$clusters, df)
}

}
