% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clusters.R
\name{extract_clusters}
\alias{extract_clusters}
\title{Extract clusters based on dissimilarity matrix}
\usage{
extract_clusters(
  diss_matrix,
  diss_threshold = 0.3,
  linkage = "single",
  vrb = TRUE
)
}
\arguments{
\item{diss_matrix}{a dissimilarity matrix.}

\item{diss_threshold}{the desired dissimilarity cutting threshold for
attributing clusters.}

\item{linkage}{the linkage function to use. This is passed down to
[fastcluster::hclust()] as parameter `method`. Please check the
documentation of [fastcluster::hclust()] for details.}

\item{vrb}{logical flag: should progress be printed to console?}
}
\value{
list object with two elements: `cl.object` (An object of class
'hclust'. It encodes a stepwise dendrogram.) and `clusters` (a data frame
with cluster memberships at the dissimilarity level `diss_threshold`).
}
\description{
This function calculates a hierarchical clustering of entries based on
a dissimilarity matrix (generated, e.g., using [calc_seq_dissimilarities()]).
It invokes [fastcluster::hclust()] internally for higher efficiency.
}
